Rename api to database
